+++ pykechain/models/sidebar/sidebar_manager.py
@@ -104,6 +104,11 @@ class SideBarManager(Iterable):
         self._perform_bulk_creation = False
         self._update()
 
+    def refresh(self) -> None:
+        """Reload the scope options from KE-chain to refresh the side-bar data."""
+        self.scope.refresh()
+        self.__init__(scope=self.scope)
+
     def remove(self, key: Any) -> None:
         """
         Remove a button from the side-bar.